+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## 3.0.0
 
-- Rate throttling is on by default.
+- Rate throttling is on by default. (https://github.com/heroku/platform-api/pull/104)
 
 ## 2.3.0
 
-- Changes are now tracked in the CHANGELOG.md
-- Rate throttling is included by default but disabled.
+- Changes are now tracked in the CHANGELOG.md.
+- Rate throttling is included by default but disabled. (https://github.com/heroku/platform-api/pull/103)